refactor: extract AnonymousQuestionSectionsList into a shared component
