+++ components/questions/AnonymousQuestionSectionsList/index.tsx
@@ -1,0 +1,3 @@
+import AnonymousQuestionSectionsList from "./AnonymousQuestionSectionsList";
+
+export default AnonymousQuestionSectionsList;
